boxin basics
The basic strategy is that every place is a list, even if there can be
only one element in the list. dndLists is used to handle the drag and
drop functionality between locations. Adding elements happens through an
in-place menu contextual to what can be placed there. The data types and
structures : variable, null, boolean, number, string, array, object,
function. Only assign and return implemented controls.