cut/paste and files
Refined a cut and paste system that doesn't rely on drag and drop. I
think this is actually better in a touch screen environement because
while drag and drop is neat, it requires accuracy, and can be tricky if
there are a lot of places it can drop nearby.

Also implementing something like a file-tree and open/close behavior.