evaluation, dnd-list eliminated, dependency file
changed the format of the evaluation box. It can also be used to add raw
javascript source code. dnd-list (drag and drop) has been removed since
I need to make it more versatile. The dependency file type was added
with the idea it would replace the method of referencing dependencies.
This would require a more custom dependency lookup system akin to npm,
which might compile to a package.json, but needs to be usable from the
front end as well.