Changing user authorization
After some thinking I have decided for now to use something similar to
json web tokens, but not store the secret on the server. That way data
on the server cannot be used to create fraudulent credentials. The token
is just a random number, hashed with a salt stored for each user.  The
hash is stored on the server and the token is sent to the client. The
token is presented as authentication by re-hashing with salt and
comparing with hash for that user in a similar way as done for
passwords, but with a less expensive algorithm.