changed file structure
I changed the structure of the view and rout files slightly to make it
more clear what things are related. Angular modules in the
public/modules folder correspond to the template files in the
views/modules folder, which are meant to go with the controllers in the
angular modules. Still trying to figure out how to properly handle
including and passing data into these templates and keeping everything
updated.