Changed authorization again
I realized what I did to replace jwt was to effectively make session
tokens that invalidated each other. Instead I changed it to generate the
same token for all logins for a given user, while still not storing
anything in the database or server which could be used to generate a
token without the users credentials.