changed template id's
Template id's don't have to be of the form of file-names with .html
extension, so I simply removed the pretense. I can't really think of a
good reason to keep the extra notation. So templates have id's with the
same value as controllers are named which are designed to go with the
template.