Examples
I have added a few example files based on a general pattern I am
developing for each type of file. So far There is a Model file for the
database, a route API file, an ejs/html template file, and an angular
module file. The first two will be general to all interfaces. The latter
two are browser specific. Each future interface target will probably
have some equivalent versions of these for the operations they will be
able to perform.