Profile view/edit/save
Implemented basics of viewing and editing elements of a profile, and
saving them to the database. searching profiles also works on all text
content of the profile. The profile data is saved as a stringified json
object which can be indexed for searches. Currently images can only be
external links.