Changed permission attributes to strings.
Decided to change attributes to strings instead of object ids of
attribute objects. A user _id also acts as an additional attribute of a
string of format "user_###" where ### is the user id, if it is added to
permissions requirements. For example, if only that attribute is added
to the 'one' requirement, then only that user would be able to perform
that action for that resource.