api security and angular
changed the way security is applied to model resources. It is now
specified either by the type of operation, or down to the route. But
security options are only added at that level if it is specified.

Also created an api factory module for angular, as well as a dependency
system, which automatically builds the required http operations to
connect to the models made by the api factory on the server.