Changed user access authorization
A guest account is now created for every visitor to the app, which can
act just like a regular account, but which cannot be logged into by
username or password, and will lose authorization at the end of a
browser session (ie close tab). But if a user logs in while the guest
account is still active, everything made with the guest account will be
transferred to the user that logs in.

There are now only two ways to grant access to a rout. Either add the
user directly to the access list of the route, or add a group to which
the user is a group to the route. The permissions can be placed either
per route, or by the type of route (safe/unsafe/update, etc), but not
both. the root access also grants full access to the resource (like the
owner option did). But there is no longer a true 'owner' of a resource,
only access permissions. The route restrictions are also available on
the client side so the client can know what can be accessed.