changed user record keeping of permissions
The models are now properties of accessRecords, and actions are now an
array, just like they are when assigning permissions. Also now using set
operations to add/remove permissions instead of search and replace. Also
added a route to the user so all their the records/permissions can be
found.