username instead of email. phone services
Uses a username instead of email address for registering a user. An
email address can now be associcated with a user using the contact
model, which is also integrating twilio as an automated phone and
messaging service. This will be a the general model for how to send
messages and notifications to people. The email is no longer a
requirement for the registration process.