Added google sign-in
Can sign-in with google+ account. This required some tricks so that
authorization through oauth2 didn't break the state of the application.
A new window is created and handed off to a url generated by the server.
(the browser never has my credentials) Google then redirects to the
callback route which renders a page which then has a piece of javascript
that calls back to the original window with the authorization code so
the original app gets the code. The code is then sent to the server
through the app in a login similar to username/password authorization
and exhanges the code for a login token generated from the oauth token.
The access_token is not currently saved anywhere. The google account
display name is used as the username, and the id is used to lookup the
same user everytime they sign-in with google.

I edited the google sign-in image that they give us. I think it is
within their regulations since the form has "sign-in with" on the left,
and maintained colors and sizes. Its just their image doesn't fit in a
button, and it wasn't transparent so it broke the overall styling
behavior of the button.