Transitioned from requireJs to webpack
I was using requireJs to handle the asynchronous loading of the single
page app. However, I have decided to try webpack. The promise is that
webpack will allow me to load modules asynchronously as the app requires
them, even template files. They are currently still loaded as before
with ejs, but my intention is to use webpack to package the template
files for async as well.