Converting to React
I am converting the front-end to use React instead of Angular. This will
involve a re-write of that part but this seems like a better way
forward.

I am also experimenting with a custom data model similar to Flux in that
it is based on actions. But I am attempting to abstract away some of the
differences between data in app vs from the rest of the internet. Redux
also looks promising as a data model if this doesn't work out.