Register/Login
re-implemented register & login in react.

Also changed the interface to the journal data model, which is more like
a messaging system. A log is now a request, and a request returns a
promise that is resolved when the requested action is completed. There
is also now a 'get' function which will only retrieve the current state
as a promise which fails if the resource does not exist, whereas a
subscription will wait even if the resource does not exist yet, until it
is manually un-subscribed.