Changes request to report
I don't want to confuse the purpose of this function. 'request' might be
confused with the request/response pattern of a normal http request.
While this function is supposed to take the place of http POST
operations, it's not the same. The action url in a report represents the
definition of the action taking place, not the data to be changed. The
state changes due to the action have to be viewed through a
subscription, or by doing a 'get'. The result of a report can be
something different than published state.