Serve webpack'd files through middleware
I took inspiration from webpack-dev-middleware. They warned not to use
that middleware for production. But that is exactly what I wanted to do.
So I looked into it and figured the only real reason I could come up
with not to use it was because it either rebuilt the bundle on every
request, or constantly monitored the files for changes to rebuild and
change the client code. I didn't care about any of that. I just want to
get rid of the extra build step, and creating additional files to keep
track of. I just want webpack to bundle the files on startup and then
serve them like a static asset. So I tried to implement that with
middleware/pack.js. Just feed in the webpack options and it builds it
and serves it from memory for the files it bundles, like it was an asset
directory. It also appears the compression also still functions
correctly for the bundled files through this middleware.