journal, apifactory, blotter
Fixes and refining of apifactory for models, and journal for front-end.
Blotter is a test model.

Models in apifactory now reflect more of the naming scheme of the
journal module. There are only actions and views. Views are what are
subscribed to. Actions are reported. Views use GET http method, and
Actions use POST, for communication between client and server. However,
this is not a REST api.

Subscriptions are not implemented yet, but when they are it will create
a symmetry between the server and client. A subscription is a way for an
action originating on one client to propagate to another client. From
the server's perspective, a view is a channel to propagate actions to
clients, through subscriptions (possibly one to many). While views can
have a REST interpretation using the GET method, which is directed by
the client, a view can also have an active nature which updates clients
directed by the server.

Also added a 'get' method to journal, which uses the same syntax as the
subscribe method, but will only update each property of a state at most
once. While a true subscription will update whenever the resources
change (when it is eventually implemented)

blotter is a test case which creates, edits a model using apifactory and
journal. It is a simple ui for adding formatted text and equations,
based on the 'profile' module that was implemented using angular.