User model is now defined using apifactory syntax.
Converted the User router and model to the new syntax, and is now
specified on loading the app. Customized user models can now be
specified by changing the user.js file path.

Also changed how api definitions are made. Now return from a
construction function instead of returning the literal object from the
model.

An api object is passed as first parameter to this function which can be
used to reference everything in the api.

Minor changes to UI, added links to blotter example