removed references to mongoose directly
since mongoose methods are referenced through the api model object now
there is really no reason to need to import mongoose into the model
definition files. The only other thing it needed was the mongoose Types
for defining the schema part, so that was also added to the api model
object.