blotter edit event pattern
reduced the amount of state being stored on each component of a blotter.
Changes are passed to the onChange callbacks which propagates the
changes up the component tree until it arrives at top, setting the new
state of the blotter, which propagates the changes back down the tree
updating the property values of the components.