editing signin setting
There are still errors when altering signin settings due to the server
authentication changing before the client authentication receives the
new values, which causes intervening requests to fail. The server
updates a user object triggering the update of that user object on the
client before the new authentication has not been updated on the clients
side.