Model Settings & Editor
Model settings can now be stored in the database, and altered by an
admin account setup the first time the app is run.

The editor function works to automate patterns that I was using, and
will hopefully help reduce errors when editing values and displaying
them with interactive ui components.