errors unsubscribing from invalid subscriptions
I think the client was thinking it made a subscription to something the
server couldn't deal with, so when it tried to unsubscribe bad things
happened since the server didn't check again. This is ok because
subscriptions can exist for things that don't, since maybe they will
exist in the future, and we want to know when they do. For now the
server just ignores those, but maybe someday it can deal with them.