split signin and registration components
Now separate components, and must switch between manually. Before it
would check if a username was registered, and would switch to the signin
mode if it was. However, this has to check with server while typing the
username which may cause delay in the correct panel displaying. and user
will probably be faster at deciding which panel they need, and may feel
light fighting the signin screen instead of an aid.