catching rendering errors
A try-catch block was added to every render method. if a render method
threw an error it seemed to leave the ui in unexpected states, and would
suppress the origin of the error which caused it, making it difficult to
know an error had occurred, and then making it more difficult to fix.

The only question was what to display when such errors happen.
Convention frowns on displaying error messages to the user since they
usually contain little to no useful information for the average user.
However, it may be more frustrating to assume that a component is not in
error state when it is. For now the standard error component is being
used.