fix report button and confirm modal
confirm modal value undefined children was converted to string instead
of going to next value option.

in report button when a progress function is used to compute the value,
start and current milliseconds are given. current ms is update every 1/4
second by the button until the report is done processing.

example function for progressValue is given in progress.jsx.